refactor(dashboard): consolidate status labels into src/lib/status-label.ts
Unify statusLabel (4 copies), displayStatus (2 copies), prettyJson (2 copies)
into a single module with comprehensive Korean status mappings.

- statusLabel: superset of all status→label mappings (30+ cases)
- displayStatus: compact command plane variant
- prettyJson: JSON pretty-printer with string pass-through

agent-roster.ts and status-badge.ts kept local functions (different semantics:
agent lifecycle labels vs general process status).

Net: -36 lines (146 removed, 110 added).
